+++ CHANGELOG.md
@@ -3,9 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.2.1] - 2019-07-15
 ### Fixed
-- Player starting health not working
+- Player starting health not working [#180]
+- Crash when dropping items
+- Crash when joining a world
+- Various other minor fixes
 
 ## [2.2.0] - 2019-07-03
 Updated for 1.14.3